lib/igt_fb: fix fb->size when provided by the user
I want to have a little more control over the size of the buffers in
kms_frontbuffer_tracking, so I decided to start calling
igt_create_fb_with_bo_size() instead of igt_create_fb(). The problem
is that create_bo_for_fb() returns its own calculated size as size_ret
instead of the actual used size.

So we fix this by returning the actual size, the one used in
gem_create instead of the calculated size that's not used anywhere.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>