kms_frontbuffer_tracking: assert the stride changes at stridechange()
We use igt_create_fb(), which decides the stride by itself: there's no
guarantee that making a buffer 512 pixels bigger is going to make its
stride change.

I had a fix for this problem that was supposed to be applied before
this patch, but due to a rework request I'm changing the order of the
patches, so we should expect to hit this assertion for now. At least
the root cause of the problem is clear now.

v2: Update the commit message due to the patch order changing.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>