scripts: remove display_debug.sh as obsolete
The script uses the obsoleted and removed intel_reg_read tool. Rather
than mechanically fix this to use intel_reg, observe that the hardcoded
register offsets are platform specific. A quick glance suggests they are
for PCH split platforms with FDI, and as such useful only on a minority
of platforms. Remove the script as obsolete.

If the need for such a script arises, it should be based around using
'intel_reg dump' with display-only register spec files.

Signed-off-by: Jani Nikula <jani.nikula@intel.com>