Revert "igt: s/basic/sanitycheck/ on prior smoketesting"
This reverts commit 4f5efc5c844f6fe69209982463f9220f8f3951ed.

There was a bit a misunderstanding on IRC between Chris&me. We want
basic tests as sanity test to be run in the BAT CI. It's just unfortunate
that right now we have fairly limited ability to absorb new ones, both
because of a pile of existing bugs in the kernel and because the CI
infrastructure is still being scaled out.

The idea was just to remove the BAT tests added yesterday, not all of
the ones we've had for a while longer.

Cc: Chris Wilson <chris@chris-wilson.co.uk>
Grumpily-acked-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>