igt/gem_softpin: Exercise snoop+uncached abutting
snooped objects are not allowed to abutt uncached objects on older gen
(!llc and global GTT) or else the GPU may hang if it prefetches across a
page boundary into a different memory type (i.e. CS reading from snoop).
The kernel should be checking the alignment rules as normal.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>