kms_frontbuffer_tracking: don't fail fbc-farfromfence on SKL
SKL doesn't do the same dspaddr_offset magic as the older gens, so FBC
may not be enabled on fbc-farfromfence. This is not a bug since FBC is
just disabled, and this is the expected case of the current Kernel, so
let's not fail the test.

If/once we land the proper infrastructure to fix this in the Kernel we
may revert this commit.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>