kms_frontbuffer_tracking: add basic subtest
This test executes one of each of the very basic operations with
whatever features are enabled by default in the Kernel, without
checking their statuses: we only care about the CRCs. Although it's
completely redundant, it will be ran by the CI so we can spot
regressions both in the specific features and in the frontbuffer
tracking subsystem.

We're not checking sink CRCs since they're still unreliable and we
don't want false negatives in the CI.

It takes 5 seconds to run this test on my machines.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>