kms_frontbuffer_tracking: don't pass the crtc idx to init_mode_params
We already pass the crtc id, so use the id to retrieve the index.
We'll change the way we pass the crtc id in the next commits, so we'll
have to call a function to calculate the index based on the id at that
point. Do the change now in order to avoid big commits later.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>